(Revert) Fully Reverted Pre-Manniversary Shortstop
Note: The shove is still kept in because a. players like it and its really useless and b. it's more trouble than its worth

The current Shortstop in the plugin appears to use the reload speed of the pre-Manniversary Shortstop:

October 13, 2011 Patch (Manniversary Update & Sale)
[Undocumented] The Shortstop reload time has been slowed down by 50%.

Historically, the pre-Manniversary Shortstop had the following stats:
-No reload speed penalty
-No push force penalty
-No bonus healing
-Shares ammo with the Scout's Pistol

Historical reference:
https://wiki.teamfortress.com/w/index.php?title=Shortstop&oldid=754125