imperfect fix for shortstop autoreload bug
fixes the autoreload bug, but then introduces another bug when autoreload is turned off. i don't think this new bug is an issue since most players have autoreload turned on.
the new bug happens when autoreload is turned off and the user presses alt-fire, which forces a reload.

this also adds in notnheavy's method in removing the shortstop shove